Carson collision metadata (#65)
* Fixed hitboxes staying on item pickup
Added methods to `Collidable` and `SpriteDict` to unregister from `SpriteDrawer`. Coupling seems tight in `Collidable` to its hitbox so this should maybe be revisited but it works for now.

* Added directional and intersection info to metadata
Directional information is calculated in `CollisionController` per collision and passed through `metadata` to collision commands. The intersection area is also passed through because that will be necessary for player-static collisions.

* Made hitbox gizmos togglable
Press G to turn hitboxes on and off.