Added oneshot animations (#83)
* Initial Implementation
Added implementation for the fundamental behavior of oneshot animations. Much refactoring still needs to be done in the classes that use `SpriteDict` to adapt to these changes and how time is delt with project-wide.

* Initial Implementation
Added implementation for the fundamental behavior of oneshot animations. Much refactoring still needs to be done in the classes that use `SpriteDict` to adapt to these changes and how time is delt with project-wide.