Adding/editing sprite sheets (#91)
Added HUD spritesheet and added doorframes to the tiles spritesheet.

Also added a new column of data to the sprite CSVs of FPS for the animations to play at. Oneshot animations that depend on animation length will be affected by changing this value. (But all unchanged sprites use the same default of 10 fps.)

For example, link now attacks with his sword faster and stays on the "use item" sprite for longer. Further fine tuning can be done by modifying the CSV files. (So long as necessary logic is using the oneshot animation system.)

NOTE: I did not update `SpriteData.cs` to include the new HUD sprite because I know there are significant pending changes in that file, so I think adding it later will be easier than introducing merge conflicts.